--- a/security.json
+++ b/security.json
@@ -1,0 +1,106 @@
+{
+  "security": {
+    "version": "2.0.0",
+    "last_updated": "2024-07-15",
+    "author": "محمدحسین نوروزی (Mohammad Hossein Norouzi)",
+    "security_headers": {
+      "content_security_policy": {
+        "enabled": true,
+        "policy": "default-src 'self'; script-src 'self' 'unsafe-inline' https://cdn.jsdelivr.net; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; img-src 'self' data: https:; font-src 'self' https://fonts.gstatic.com; connect-src 'self' https:; frame-src 'none'; object-src 'none'",
+        "description": "Prevents XSS attacks and unauthorized resource loading"
+      },
+      "x_content_type_options": {
+        "enabled": true,
+        "value": "nosniff",
+        "description": "Prevents MIME type sniffing"
+      },
+      "x_frame_options": {
+        "enabled": true,
+        "value": "DENY",
+        "description": "Prevents clickjacking attacks"
+      },
+      "x_xss_protection": {
+        "enabled": true,
+        "value": "1; mode=block",
+        "description": "Enables XSS filtering"
+      },
+      "referrer_policy": {
+        "enabled": true,
+        "value": "strict-origin-when-cross-origin",
+        "description": "Controls referrer information"
+      },
+      "permissions_policy": {
+        "enabled": true,
+        "value": "geolocation=(), microphone=(), camera=(), payment=(), usb=()",
+        "description": "Restricts dangerous browser features"
+      }
+    },
+    "https_enforcement": {
+      "enabled": true,
+      "hsts_max_age": 31536000,
+      "include_subdomains": true,
+      "preload": true
+    },
+    "input_validation": {
+      "form_validation": true,
+      "sanitization": true,
+      "csrf_protection": true
+    },
+    "security_monitoring": {
+      "error_logging": true,
+      "security_alerts": true,
+      "audit_trail": true
+    }
+  },
+  "compliance": {
+    "gdpr": {
+      "data_protection": true,
+      "privacy_policy": true,
+      "consent_management": true
+    },
+    "accessibility": {
+      "wcag_2_1": true,
+      "screen_reader_support": true,
+      "keyboard_navigation": true
+    },
+    "seo": {
+      "structured_data": true,
+      "meta_tags": true,
+      "sitemap": true
+    }
+  },
+  "threat_model": {
+    "common_attacks": [
+      "XSS (Cross-Site Scripting)",
+      "CSRF (Cross-Site Request Forgery)",
+      "Clickjacking",
+      "MIME Type Sniffing",
+      "Content Injection",
+      "Data Leakage"
+    ],
+    "mitigations": [
+      "Content Security Policy",
+      "Input Validation",
+      "Output Encoding",
+      "Secure Headers",
+      "HTTPS Enforcement",
+      "Regular Security Updates"
+    ]
+  },
+  "recommendations": {
+    "server_configuration": [
+      "Enable HTTPS/TLS 1.3",
+      "Configure secure headers",
+      "Implement rate limiting",
+      "Regular security updates",
+      "Monitor security logs"
+    ],
+    "code_security": [
+      "Validate all user inputs",
+      "Sanitize output data",
+      "Use parameterized queries",
+      "Implement proper error handling",
+      "Regular security testing"
+    ]
+  }
+}